seccon/14/chamber-of-echos/requirements.txt
pycryptodome==3.23.0
scapy==2.6.1
